Working on finding leftmost child
I added some more comments of though process and purpose of method. I set up a loop to find the lowest non-leaf node that has children starting at the desired range. I'm not sure if the bounds are quite correct, but I went off of the B+ tree slide structure.