Added 0 Key value check to startScan should be done
If the key value [i] is supposed to be 0, then there will be a  valid page number at i+1 in the pageNoArry. If the key is supposed to be null there won't.

If the key is null we know that the page number in the i index will be the range we are looking for as this is the range that is larger than the last valid key value